requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
distlib==0.3.6
django-allauth==0.54.0
django-filter==23.1
django-rest-auth==0.9.5
Django==3.2.5
djangorestframework-simplejwt==5.2.2
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.11.0
gunicorn==20.1.0
idna==3.4
Markdown==3.4.3
oauthlib==3.2.2
openpyxl==3.1.2
pip==23.1.2
pipenv==2023.3.20
platformdirs==3.2.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
requests==2.29.0
setuptools==67.6.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
virtualenv-clone==0.5.7
virtualenv==20.21.0
